+++ app/views/dashboard/manager.html.slim
