+++ app/views/dashboard/report_sales.html.slim
@@ -1,0 +1,38 @@
+head
+  title Sales Up - Sales Report
+body
+  .row
+    .panel.panel-default style="margin-top: -30px;"
+      .panel-heading
+        .form-inline
+          h3.h3.form-group
+            = "Name:"
+          h3.h3.font-thin.form-group.m-l = @current_report.report_name
+        .form-inline
+          h3.h3.form-group
+            = "Goal: "
+          h3.h3.font-thin.form-group.m-l = "R" + number_to_currency(@current_report.goal).to_s
+
+      .panel-body style="font-size: 25px; font-weight: bold;"
+        div id="placeholderG" style="height:450px;width:1250px;"
+
+      div.panel-default style="margin-top: 15px;"
+        table.table ui-jq="footable"
+          thead
+            tr
+              th Salesman Name
+              th Contracts Closed
+              th Contract Sum
+              th Closed Per Active Day
+              th Goal's Percentage
+              th Options
+          tbody
+            - @record_data.each_with_index do |data, i|
+              tr
+                td = data[0]
+                td = data[1]
+                td = "R" + number_to_currency(data[2]).to_s
+                td = data[3]
+                td = data[4].to_s + "%"
+                td
+
