+++ app/views/layouts/menu.html.slim
@@ -18,7 +18,7 @@ html
                   br>/
                   = current_user.full_name
             .form-group#menu-search
-              - unless current_page?(controller: 'page', action: 'overview') || current_page?(controller: 'page', action: 'manager')
+              - unless current_page?(controller: 'dashboard', action: 'overview') || current_page?(controller: 'dashboard', action: 'manager')
                 = simple_form_for @report, method: :get, url: "/monthly_sales", html: { class: "form-group", id: "formSelectBox" } do |f|
                   = f.input :month, as: :hidden, input_html: { value: "p", id: "monthSelected" }
                   = f.input :year, label: false, :collection => @month_year_list.map { |report| [ report.month+"/"+report.year.to_s, report.month+"/"+report.year.to_s] }, :prompt => false, input_html: { class: "form-control", id: "monthBox",  :onchange => "sendData('formSelectBox', 0)"}
@@ -29,26 +29,32 @@ html
               button.btn.btn-primary.btn-addon.btn-lg.margin.dropdown-toggle data-toggle="dropdown" id="menu1" Monthly
               ul.dropdown-menu.dropdown-li role="menu" aria-labelledby="menu1"
                 li.dropdown-li role="presentation"
-                  = simple_form_for @report, method: :get, url: "/report_sales", html: { class: "form-group", style: "margin-bottom: 8px; margin-left: 1px;" } do |f|
-                    - unless params[:report].nil? || params[:report][:month].blank? || params[:report][:month].include?("/")
-                      = f.input :month, as: :hidden, input_html: { value: params[:report][:month] }
-                      = f.input :year, as: :hidden, input_html: { value: params[:report][:year] }
-                    = f.button :submit, "Reports", class: "btn btn-primary btn-addon btn-lg margin"
+                  - if current_page?(controller: 'dashboard', action: 'report_sales')
+                    button.btn.btn-primary.btn-addon.btn-lg.disabled.margin style="margin-bottom: 8px; margin-left: 6px;" Reports
+                  - else
+                    = simple_form_for @report, method: :get, url: "/report_sales", html: { class: "form-group", style: "margin-bottom: 8px; margin-left: 1px;" } do |f|
+                      - unless params[:report].nil? || params[:report][:month].blank? || params[:report][:month].include?("/")
+                        = f.input :month, as: :hidden, input_html: { value: params[:report][:month] }
+                        = f.input :year, as: :hidden, input_html: { value: params[:report][:year] }
+                      = f.button :submit, "Reports", class: "btn btn-primary btn-addon btn-lg margin"
                 li.dropdown-li role="presentation"
-                  = simple_form_for @report, method: :get, url: "/monthly_sales", html: { class: "form-group", style: "margin-left: 10px;" } do |f|
-                    - unless params[:report].nil? || params[:report][:month].blank? || params[:report][:month].include?("/")
-                      = f.input :month, as: :hidden, input_html: { value: params[:report][:month] }
-                      = f.input :year, as: :hidden, input_html: { value: params[:report][:year] }
-                    = f.button :submit, "Sales", class: "btn btn-primary btn-addon btn-lg margin"
+                  - if current_page?(controller: 'dashboard', action: 'monthly_sales')
+                    button.btn.btn-primary.btn-addon.btn-lg.disabled.margin style="margin-left: 15px;" Sales
+                  - else
+                    = simple_form_for @report, method: :get, url: "/monthly_sales", html: { class: "form-group", style: "margin-left: 10px;" } do |f|
+                      - unless params[:report].nil? || params[:report][:month].blank? || params[:report][:month].include?("/")
+                        = f.input :month, as: :hidden, input_html: { value: params[:report][:month] }
+                        = f.input :year, as: :hidden, input_html: { value: params[:report][:year] }
+                      = f.button :submit, "Sales", class: "btn btn-primary btn-addon btn-lg margin"
 
-              - if current_page?(controller: 'page', action: 'overview')
+              - if current_page?(controller: 'dashboard', action: 'overview')
                 button.btn.btn-primary.btn-addon.btn-lg.disabled.margin Overview
               - else
                 = simple_form_for @report, method: :get, url: "/overview", html: { class: "form-group" } do |f|
                   = f.button :submit, "Overview", class: "btn btn-primary btn-addon btn-lg margin"
 
-              - if can? :manager, PageController
-                - if current_page?(controller: 'page', action: 'manager')
+              - if can? :manager, DashboardController
+                - if current_page?(controller: 'dashboard', action: 'manager')
                   button.btn.btn-primary.btn-addon.btn-lg.disabled.margin Settings
                 - else
                   = simple_form_for @report, method: :get, url: "/manager", html: { class: "form-group" } do |f|
