--- app/views/page/report_sales.html.slim
@@ -1,3 +1,0 @@
-head
-  title Sales Up - Sales Report
-body
