+++ config/routes.rb
@@ -5,11 +5,9 @@ Rails.application.routes.draw do
   root to: 'user#user_home'
   get '/forgot_password', to: 'user#forgot_password'
 
-  get '/monthly_sales', to: 'page#monthly_sales'
-  get '/report_sales', to: 'page#report_sales'
-  get '/manager', to: 'page#manager'
-  get '/overview', to: 'page#overview'
-  get '/logout', to: 'page#logout'
-
-  get '/search', to: 'dashboard#search'
+  get '/monthly_sales', to: 'dashboard#monthly_sales'
+  get '/report_sales', to: 'dashboard#report_sales'
+  get '/manager', to: 'dashboard#manager'
+  get '/overview', to: 'dashboard#overview'
+  get '/logout', to: 'dashboard#logout'
 end
